BUG #2638 - fix bug that keep slow log on, dble may occur oom
