+++ src/main/java/com/actiontech/dble/route/impl/AbstractRouteStrategy.java
@@ -13,21 +13,48 @@ import com.actiontech.dble.route.util.RouterUtil;
 import com.actiontech.dble.server.ServerConnection;
 import com.actiontech.dble.server.parser.ServerParse;
 import com.actiontech.dble.sqlengine.mpp.LoadData;
+import com.alibaba.druid.sql.ast.SQLStatement;
+import com.alibaba.druid.sql.dialect.mysql.parser.MySqlStatementParser;
+import com.alibaba.druid.sql.parser.SQLStatementParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
+import java.sql.SQLSyntaxErrorException;
+import java.util.List;
 
 public abstract class AbstractRouteStrategy implements RouteStrategy {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractRouteStrategy.class);
 
     @Override
+    public SQLStatement parserSQL(String originSql) throws SQLSyntaxErrorException {
+        SQLStatementParser parser = new MySqlStatementParser(originSql);
+
+        /**
+         * thrown SQL SyntaxError if parser error
+         */
+        try {
+            List<SQLStatement> list = parser.parseStatementList();
+            if (list.size() > 1) {
+                throw new SQLSyntaxErrorException("MultiQueries is not supported,use single query instead ");
+            }
+            return list.get(0);
+        } catch (Exception t) {
+            LOGGER.warn("routeNormalSqlWithAST", t);
+            if (t.getMessage() != null) {
+                throw new SQLSyntaxErrorException(t.getMessage());
+            } else {
+                throw new SQLSyntaxErrorException(t);
+            }
+        }
+    }
+
+    @Override
     public RouteResultset route(SchemaConfig schema, int sqlType, String origSQL,
                                 ServerConnection sc, LayerCachePool cachePool) throws SQLException {
 
         RouteResultset rrs = new RouteResultset(origSQL, sqlType);
-
 
         /*
          * debug mode and load data ,no cache
@@ -50,7 +77,7 @@ public abstract class AbstractRouteStrategy implements RouteStrategy {
                 rrs = routeNormalSqlWithAST(schema, origSQL, rrs, cachePool, sc);
             }
         }
-
+        sc.getSession2().endParse();
         return rrs;
     }
 
@@ -58,8 +85,8 @@ public abstract class AbstractRouteStrategy implements RouteStrategy {
     /**
      * routeNormalSqlWithAST
      */
-    public abstract RouteResultset routeNormalSqlWithAST(SchemaConfig schema, String stmt, RouteResultset rrs,
-                                                         LayerCachePool cachePool, ServerConnection sc) throws SQLException;
+    protected abstract RouteResultset routeNormalSqlWithAST(SchemaConfig schema, String stmt, RouteResultset rrs,
+                                                            LayerCachePool cachePool, ServerConnection sc) throws SQLException;
 
 
 }
