+++ src/main/java/com/actiontech/dble/route/impl/DefaultRouteStrategy.java
@@ -21,65 +21,37 @@ import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
 import java.sql.SQLSyntaxErrorException;
-import java.util.List;
 
 public class DefaultRouteStrategy extends AbstractRouteStrategy {
 
-    public static final Logger LOGGER = LoggerFactory.getLogger(DefaultRouteStrategy.class);
-
-
-    public SQLStatement parserSQL(String originSql, ServerConnection c) throws SQLSyntaxErrorException {
-        SQLStatementParser parser;
-        parser = new MySqlStatementParser(originSql);
-        try {
-            return parser.parseStatement(true);
-        } catch (Exception t) {
-            LOGGER.info("routeNormalSqlWithAST", t);
-            if (t.getMessage() != null) {
-                throw new SQLSyntaxErrorException(t.getMessage());
-            } else {
-                throw new SQLSyntaxErrorException(t);
-            }
-        }
-    }
+    private static final Logger LOGGER = LoggerFactory.getLogger(DefaultRouteStrategy.class);
 
     @Override
-    public SQLStatement parserSQL(String originSql) throws SQLSyntaxErrorException {
-        SQLStatementParser parser = new MySqlStatementParser(originSql);
-
-        /**
-         * thrown SQL SyntaxError if parser error
-         */
-        try {
-            List<SQLStatement> list = parser.parseStatementList();
-            if (list.size() > 1) {
-                throw new SQLSyntaxErrorException("MultiQueries is not supported,use single query instead ");
-            }
-            return list.get(0);
-        } catch (Exception t) {
-            LOGGER.info("routeNormalSqlWithAST", t);
-            if (t.getMessage() != null) {
-                throw new SQLSyntaxErrorException(t.getMessage());
-            } else {
-                throw new SQLSyntaxErrorException(t);
-            }
-        }
-    }
-
-    @Override
-    public RouteResultset routeNormalSqlWithAST(SchemaConfig schema,
-                                                String originSql, RouteResultset rrs,
-                                                LayerCachePool cachePool, ServerConnection sc) throws SQLException {
-        SQLStatement statement = parserSQL(originSql, sc);
+    protected RouteResultset routeNormalSqlWithAST(SchemaConfig schema,
+                                                   String originSql, RouteResultset rrs,
+                                                   LayerCachePool cachePool, ServerConnection sc) throws SQLException {
+        SQLStatement statement = parse(originSql);
         if (sc.getSession2().getIsMultiStatement().get()) {
             originSql = statement.toString();
             rrs.setStatement(originSql);
             rrs.setSrcStatement(originSql);
         }
-        sc.getSession2().endParse();
         DruidParser druidParser = DruidParserFactory.create(statement, rrs.getSqlType());
         return RouterUtil.routeFromParser(druidParser, schema, rrs, statement, originSql, cachePool, new ServerSchemaStatVisitor(), sc, null);
+    }
 
+    private SQLStatement parse(String originSql) throws SQLSyntaxErrorException {
+        SQLStatementParser parser = new MySqlStatementParser(originSql);
+        try {
+            return parser.parseStatement(true);
+        } catch (Exception t) {
+            LOGGER.warn("routeNormalSqlWithAST", t);
+            if (t.getMessage() != null) {
+                throw new SQLSyntaxErrorException(t.getMessage());
+            } else {
+                throw new SQLSyntaxErrorException(t);
+            }
+        }
     }
 
 }
