+++ src/main/java/com/actiontech/dble/server/handler/ServerLoadDataInfileHandler.java
@@ -199,6 +199,7 @@ public final class ServerLoadDataInfileHandler implements LoadDataInfileHandler 
                     if (rrs != null) {
                         flushDataToFile();
                         isStartLoadData = false;
+                        serverConnection.getSession2().endParse();
                         serverConnection.getSession2().execute(rrs);
                     }
                 }
@@ -626,6 +627,7 @@ public final class ServerLoadDataInfileHandler implements LoadDataInfileHandler 
         RouteResultset rrs = buildResultSet(routeResultMap);
         if (rrs != null) {
             flushDataToFile();
+            serverConnection.getSession2().endParse();
             serverConnection.getSession2().execute(rrs);
         }
     }
